pyproject.toml
[project]
name = "image-deduplicator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "imagehash>=4.3.2",
    "opencv-python>=4.13.0.92",
    "pillow>=12.1.1",
    "pillow-heif>=1.3.0",
    "pytest>=9.0.2",
    "pytest-xdist>=3.8.0",
    "rich>=14.3.3",
    "typer>=0.24.1",
]
